Summarize Market Trends by Year

Aggregates annual statistics, such as average median price, average growth rate, and the number of localities tracked, by year.

market_trends_by_year(data)

Arguments

data

Data frame containing property data with columns 'year', 'median_price', and 'Locality'.

Value

A data frame with columns 'year', 'avg_median_price', 'avg_growth_rate', and 'locality_count', summarizing trends per year.

Examples

market_trends_by_year(property_data)
#>    year avg_median_price avg_growth_rate locality_count
#> 1  2013         477691.0      -0.1655769            787
#> 2  2014         512858.8      -0.1644211            787
#> 3  2015         571694.2      -0.2083496            787
#> 4  2016         625686.9      -0.2104521            787
#> 5  2017         704482.1      -0.1987476            787
#> 6  2018         716641.4      -0.2018557            787
#> 7  2019         704250.3      -0.1805207            787
#> 8  2020         758757.3      -0.1795944            787
#> 9  2021         902710.9      -0.1370067            787
#> 10 2022         948135.3      -0.1137488            787
#> 11 2023         923986.7      -0.1004654            787